pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "git-alchemist"
version = "1.2.2"
description = "A unified AI stack to optimize, describe, architect, and forge pull requests for your GitHub repositories."
readme = "README.md"
authors = [{ name = "abduznik" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "rich",
    "python-dotenv",
    "requests"
]

[project.scripts]
alchemist = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
